configure.ac: add --disable-gtk4 flag.
Previously, the HAVE_GTK_4_0 macro has been enabled whenever GTK 4
development files were found on the system.

We keep that behavior, and actually modify the build system to actually
build and link against GTK 4, but also add a new configure flag
--disable-gtk4 that disables GTK 4 usage.

The logic goes like that:
  - If --disable-gtk4 has not been provided, configure will first look
    for GTK 4 and use that if available, otherwise fall back to GTK+ 3.
  - If --disable-gtk4 has been provided, GTK 4 detection is skipped and
    only GTK+ 3 is detected.
  - If --enable-gtk4 has been provided, configure will check for GTK 4
    and not fall back to GTK+ 3, but error out if GTK 4 has not been
    found.

Note that making GTK 4 the default might not be the best option
currently, because the greeter does not build against GTK 4, but is
paving the way for the future.

For now, users will have to use --disable-gtk4, up until all code has
been fully ported to support GTK 4.

Fixes: https://github.com/ArcticaProject/arctica-greeter/issues/56